Restructured test separating internal and external test
Previously all test could access the internal functionality of the library. Tests
have now been separated into internal and external, this allows external test to
be executed as if they were a user of the library. And internal test can continue
to run, testing the unexported functionality of the library.